% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseThemeGMR.R
\name{BaseThemeGMR}
\alias{BaseThemeGMR}
\title{Base theme for plotting with gmr}
\usage{
BaseThemeGMR(
  leg.pos = "bottom",
  base_size = 14,
  x_angle = 0,
  x_axis_size = 10,
  y_axis_size = 10,
  do.axis.line = TRUE,
  axis_title_size = 12,
  do.ticks = TRUE,
  do.grid = TRUE,
  strip_text_size = 12,
  panel_space_x = 1.5,
  panel_space_y = 1.5,
  panel_face = "plain",
  plot_title_size = 16,
  plot_title_margin = 10,
  subtitle_size = 12,
  subtitle_margin = 12,
  subtitle_face = "plain",
  caption_size = 9,
  caption_margin = 10,
  tag_size = 12,
  plot_margin = ggplot2::margin(30, 30, 30, 30),
  ...
)
}
\description{
Base theme for plotting with gmr
}
