% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_files.R
\name{clean_bat_Sim}
\alias{clean_bat_Sim}
\title{Clean the output files of the Gmacs simulation module}
\usage{
clean_bat_Sim(path = ".", verbose = TRUE)
}
\arguments{
\item{path}{(character string)- names or file path representing root (default: ".")}

\item{verbose}{(TRUE/FALSE) flag to print file names being deleted}
}
\value{
nothing
}
\description{
Function to clean gmacs output files in a folder when a
simulation-estimation approach has been considered.
}
\details{
Deletes files matching \code{"admodel.\*"},\code{"\*.bar"},\code{"\*.eva"},
\code{"fmin.log"},\code{"\*.log"},\code{"\*.std"},\code{"gradient.\*"},
\code{"\*.r0\*"},\code{"\*.p0\*"},\code{"\*.b0\*"},\code{"derivatives"},
\code{"ders.dat"},\code{"mcout\*.\*"},\code{"checkfile.rep"},\code{"personal.rep"},
\code{"gmacs_in.\*"},\code{"gmacs_files_in.dat"},\code{"\*.exe"},\code{"\*.a"},
\code{"gmacs.par"},\code{"gmacs.dat"},\code{"gmacs.rep"},\code{"simdata.out"}.
}
\seealso{
[clean_files()]
}
