% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_files.R
\name{clean_files}
\alias{clean_files}
\title{Clean files in a folder}
\usage{
clean_files(path = ".", names = c("no_file"), verbose = FALSE)
}
\arguments{
\item{path}{character string or file path representing the root folder to be cleaned}

\item{names}{vector of character strings representing files in the root folder to be removed}

\item{verbose}{(TRUE/FALSE) flag to print filenames being deleted}
}
\value{
nothing
}
\description{
Function to clean files in a folder.
}
\details{
removes (deletes) files with given names on given path.
}
