% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_files.R
\name{clean_Inputfiles}
\alias{clean_Inputfiles}
\title{Clean Gmacs input files in a folder}
\usage{
clean_Inputfiles(path = ".", verbose = FALSE)
}
\arguments{
\item{path}{(character string; path)- the root folder to be cleaned}

\item{verbose}{(logical)- flag to print filenames being deleted}
}
\value{
nothing
}
\description{
Function to clean Gmacs input files in a specifc folder.
}
\details{
removes (deletes) Gmacs input files based on those available in the
'gmacs.dat' file.
}
