% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_files.R
\name{clean_root}
\alias{clean_root}
\title{Clean gmacs compilation-process files in the "root folder"}
\usage{
clean_root(path = ".", verbose = TRUE)
}
\arguments{
\item{path}{character string or file path representing root (default: ".")}

\item{verbose}{(TRUE/FALSE) flag to print file names being deleted}
}
\value{
nothing
}
\description{
Function to clean gmacs compilation-process files in the "root folder".
}
\details{

}
