% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copy_GMACS_Input_files.R
\name{copy_GMACSinputs}
\alias{copy_GMACSinputs}
\title{Copy specific GMACS input files}
\usage{
copy_GMACSinputs(
  fromDir = NULL,
  toDir = NULL,
  GMACS_files = NULL,
  overwrite = NULL
)
}
\arguments{
\item{fromDir}{(path)- path to the directory where the input files to copy
are stored.}

\item{toDir}{(path)- path to the directory where to copy the files.}

\item{GMACS_files}{(character string)- name of the GMACS input file(s) you want to
copy. Options for \code{GMACS_files} are: 'gmacs.dat','gmacs.pin',
the names of either the control, data, or projection files
(with extention files, i.e.: .ctl, .dat, .prj).
If \code{GMACS_files="all"}, then the function will copy
and paste all the input files cited above.
from the base model directory to your current model in development.}

\item{overwrite}{(logical) - Overwrite the input files if they already exist
in the \code{toDir} repertory.}
}
\description{
Reads the gmacs.dat file to figure out the names of the control, data,
and projection files, then copies those files along with gmacs.dat and
gmacs.pin (if applicable).
}
\author{
Matthieu Veron
}
