% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGMACS.R
\name{.buildGMACS}
\alias{.buildGMACS}
\title{.buildGMACS}
\usage{
.buildGMACS(
  prefix = NULL,
  raneff = NULL,
  safe = NULL,
  dll = NULL,
  debug = NULL,
  logfile = NULL,
  add = NULL,
  verbose = NULL,
  pathfile = NULL,
  args = NULL
)
}
\arguments{
\item{prefix}{String name prefix (i.e., without extension) of the ADMB project;
here "gmacs".}

\item{raneff}{logical: if \code{TRUE}, use the random effects model executable
    \code{tpl2rem.exe}, otherwise use the normal model executable \code{tpl2cpp.exe}.}

\item{safe}{logical: if \code{TRUE}, use safe mode with bounds checking on all
    array objects, otherwise use optimized mode for fastest execution.}

\item{dll}{create dll (rather than executable)}

\item{debug}{compile with debug symbols}

\item{logfile}{logical: if \code{TRUE}, create a log file of the messages from the shell call.}

\item{add}{logical: if \code{TRUE}, append shell call messages to an exsiting log file.}

\item{verbose}{logical: if \code{TRUE}, report the shell call an its messages to the R console.}

\item{pathfile}{string name of 2-column text file that details the 
    relevant paths for the R variables \code{admbpath}, \code{gccpath}, and \code{editor}.}

\item{args}{list of string names corresponding to the various .cpp libraries
required to build gmacs. These files are stored in the lib/ folder.}
}
\value{
nothing.
}
\description{
Function to build the GMACS executable.
It converts the gmacs.tpl code to a gmacs.cpp code, calls ADMB to compile all
 .cpp files required to build gmacs and returns the gmacs executable.
}
