% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMACS_run_Utilities.R
\name{.CallTerm}
\alias{.CallTerm}
\title{.CallTerm}
\usage{
.CallTerm(command = command, .Dir = NULL, verbose = NULL)
}
\arguments{
\item{command}{Character string: name of the command to run (.exe)}

\item{.Dir}{Character string: the directory of the executable}

\item{verbose}{Logical: if TRUE, report the shell call an its messages 
to the R console.}
}
\value{
the (unique) terminal identifier
}
\description{
Calls a R terminal to run a specific executable
}
\examples{
\dontrun{
}


}
