% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmr-options.R
\docType{data}
\name{.cbPalette1}
\alias{.cbPalette1}
\title{The color-blind friendly palette with grey}
\format{
An object of class \code{character} of length 8.
}
\usage{
.cbPalette1
}
\description{
see http://www.cookbook-r.com/Graphs/Colors_%28ggplot2%29/
}
\author{
D'Arcy N. Webber
}
\keyword{datasets}
