% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dynB0.R
\name{.get_dynB0_df}
\alias{.get_dynB0_df}
\title{Extract spawning stock biomass (ssb) from gmacs run}
\usage{
.get_dynB0_df(M)
}
\arguments{
\item{M}{list object created by read_admb function}
}
\value{
dataframe of spawning biomass
}
\description{
Spawning biomass may be defined as all males or some combination of males and
females
}
