% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-F.R
\name{.get_F_df}
\alias{.get_F_df}
\title{Get F parameters}
\usage{
.get_F_df(M)
}
\arguments{
\item{M}{List object(s) created by read_admb function}
}
\value{
dataframe of Fs
}
\description{
Get F parameters
}
\author{
D'Arcy N. Webber
}
