% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-growth-tag.R
\name{.get_gi_tag_df}
\alias{.get_gi_tag_df}
\title{Get growth increments from tagging data}
\usage{
.get_gi_tag_df(M)
}
\arguments{
\item{M}{list object(s) created by read_admb function}
}
\value{
dataframe of growth increments
}
\description{
Get growth increments from tagging data
}
\author{
D'Arcy N. Webber, SJD Martell
}
