% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-meanSize.R
\name{.get_mnLen_df}
\alias{.get_mnLen_df}
\title{Get observed and predicted mean lengths}
\usage{
.get_mnLen_df(M)
}
\arguments{
\item{M}{List object(s) created by read_admb function}
}
\value{
dataframe of catch history (observed) and predicted values
}
\description{
Get observed and predicted mean lengths
}
\author{
J. Ianelli, SJD Martell, D'Arcy N. Webber
}
