% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVerGMACS.R
\name{.getVerGMACS}
\alias{.getVerGMACS}
\title{Give a new version name}
\usage{
.getVerGMACS()
}
\value{
Character string: the user-defined name of the new GMACS version
}
\description{
Function to ask the user the name of the new Version of GMACS
and specify the compilation date for the new gmacbase.TPL file when updating it.
}
