% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_module_functions.R
\name{Gen_GmacsSim}
\alias{Gen_GmacsSim}
\title{Generates the simulated data for a set of simulation}
\usage{
Gen_GmacsSim(
  path = NULL,
  Stock = NULL,
  RandomRec = NULL,
  Ndev_Rec = NULL,
  SigmaR = NULL,
  RandSim = NULL,
  UseParam = NULL,
  No_Est = NULL,
  Nsim = NULL,
  seed_val = NULL,
  verbose = NULL
)
}
\arguments{
\item{path}{(character string)- file path representing root i.e., the folder where
the \code{gmacsbase.tpl} file is stored.}

\item{Stock}{(character string)- Name of the stock of interest. This will be used
to set the working directories: this is the name of the folder in the
\code{build} directory where the basic gmacs input files are stored.}

\item{RandomRec}{(logical)- if \code{FALSE} : add deviations to the original ones,
if \code{TRUE} generate new deviations.}

\item{Ndev_Rec}{(integer)- Number of year to which apply a deviation in recruitment.}

\item{SigmaR}{(numeric)- Standard deviation used in the \code{\link[stats:Normal]{stats::rnorm()}} function.}

\item{RandSim}{(logical)- If true, random deviations are generated.}

\item{UseParam}{(logical)- Do the simulations use a .PIN file to import parameters?}

\item{No_Est}{(logical)- Set the maximum function call to 1? i.e., run estimation?}

\item{Nsim}{(integer)- the number of simulation to run.}

\item{seed_val}{(integer) - a random number generation (single value; see \code{\link[base:Random]{base::set.seed()}}).}

\item{verbose}{(logical)- flag to print processing information}
}
\value{
Nothing
}
\description{
Function that generates the simulated data for a set of simulation
with Gmacs.
}
