% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_module_functions.R
\name{GenSimFiles}
\alias{GenSimFiles}
\title{Generate the input files to get the simulated data for each simulation}
\usage{
GenSimFiles(
  Isim = NULL,
  OrigFiles = NULL,
  Ndev_Rec = NULL,
  SigmaR = NULL,
  DoRecDev = NULL,
  verbose = NULL,
  FunCall = TRUE
)
}
\arguments{
\item{Isim}{(integer)- simulation number (1 to Nsim)}

\item{OrigFiles}{(named list)- Output of the \code{\link[=prepSim]{prepSim()}} function.}

\item{Ndev_Rec}{(integer)- Number of year to which apply a deviation in recruitment}

\item{SigmaR}{(numeric)- Standard deviation used in the \code{\link[stats:Normal]{stats::rnorm()}} function}

\item{DoRecDev}{(logical)- if \code{FALSE} : add deviation to the original ones,
if \code{TRUE} generate new deviations.}

\item{verbose}{(logical)- flag to print processing information.
to generate random deviations that will be added to the recruitment deviation.}

\item{FunCall}{(logical)- Run the executable (with no estimation) ?}
}
\value{
the SimData.out file as a named list for each simulation.
}
\description{
Generates the input files for each of the Nsim simulations.
Specifically, this function re-write the original input files in the \code{True} folder
and:
\enumerate{
\item allows for considering random variation in recruitment (either by adding random
variations (normally distributed) to the original recruitment deviations or by
generating "new deviations",
\item Set the maximum number of function calls to 1,
\item call the gmacs.exe (with no estimation) to generate a new gmacsall.out file
that will be used for the simulations.
}

The function is called for each simulation so it will generate Nsim gmacsall.out
which are copy and paste in the \code{Run} folder.
}
\seealso{
\code{\link{readGMACSfiles}}, \code{\link{writeGmacs.dat}},
\code{\link{readGMACSdat}},\code{\link{writeGmacsdatfile}},
\code{\link{readGMACSctl}}, \code{\link{writeGmacsctlfile}},
\code{\link{readGMACSprj}}, \code{\link{writeGmacsprjfile}},
\code{\link{readGMACSpar}},\code{\link{writeGmacsPAR}},
\code{\link{readGMACSsimdat}},
}
