% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Param_name.R
\name{get_Param_name}
\alias{get_Param_name}
\title{Get the name of all parameters}
\usage{
get_Param_name(DatFile = NULL, CtlFile = NULL, nyrRetro = NULL)
}
\arguments{
\item{DatFile}{(list)- Object containing the .dat file - This is the output
of the \code{\link[=readGMACSdat]{readGMACSdat()}} function.}

\item{CtlFile}{(list)- Object containing the .ctl file - This is the output
of the \code{\link[=readGMACSctl]{readGMACSctl()}} function.}

\item{nyrRetro}{(integer)- Number of year for the retrospective analysis}
}
\value{
the name as the parameters as a named list.
}
\description{
This functions returns the name of each parameter based on the
model specification (sexes, maturity, fleet, ...).
}
\seealso{
\code{\link{readGMACSallOUT}}.
}
