% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInpOutFiles.R
\name{getInpOutFiles}
\alias{getInpOutFiles}
\title{Read all input and output files}
\usage{
getInpOutFiles(Dir = NULL, verbose = NULL)
}
\arguments{
\item{Dir}{(character string)- path where the input/out files are saved for the
stock of interest.}
}
\value{
a named list of lists:
\describe{
\item{\code{GMACSdat}}{the \code{gmacs.dat} file;}
\item{\code{datFile}}{the \code{model.dat} file;}
\item{\code{ctlFile}}{the \code{model.ctl} file;}
\item{\code{prjfile}}{the \code{model.prj} file;}
\item{\code{GmacsAll_out}}{the \code{Gmacsall.out} file if available;}
\item{\code{gmacs_par}}{the \code{gmacs.par} file if available;}
\item{\code{gmacs_rep}}{the \code{gmacs.rep} file if available;}
\item{\code{gmacs_sim}}{the \code{simdata.out} file if available}
}
}
\description{
This function reads all the Gmacs input files (\code{"gmacs.dat"},
\code{"model.dat"}, \code{"model.ctl"}, \code{"model.prj"} and if they exist, all the
output files (\code{"Gmacsall.out"}, \code{"gmacs.par"}, \code{"gmacs.rep"},
\code{"simdata.out"})).
}
\examples{
\dontrun{
# Load package ----
library(gmr)

# Set directories ----
Dir_Dvpt_Vers <- file.path(here::here(), "Dvpt_Version", fsep = fsep)
Dir_Last_Vers <- file.path(here::here(), "Latest_Version", fsep = fsep)

# Define the stock of interest ----
stock <- "SMBKC"

# Read Gmacs files ----
Stock_files <-
getInpOutFiles(Dir = file.path(Dir_Dvpt_Vers, 'build', stock, fsep = fsep),
verbose = TRUE)
}

}
\seealso{
\code{\link{readGMACS.dat}},\code{\link{readGMACSdat}},
\code{\link{readGMACSctl}}, \code{\link{readGMACSprj}}.
}
