% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVerGMACS.R
\name{GetVerSpec}
\alias{GetVerSpec}
\title{Get the specifications of the latest version}
\usage{
GetVerSpec(Dir = NULL)
}
\arguments{
\item{Dir}{(character string)- path of the folder name that holds the version of
GMACS that is updated.}
}
\value{
Creates the GMACS_Version_details.txt which is store in the Latest_version
folder.
}
\description{
This function creates the GMACS_Version_details.txt file holding
that holds the specifications of the latest version of GMACS. This function
is called by the UpdateGMACS() function when a developer wants to udpate GMACS
before submitting a pull request for a potential new release.
}
