% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMACS.R
\name{GMACS}
\alias{GMACS}
\title{GMACS}
\usage{
GMACS(
  Spc = NULL,
  GMACS_version = NULL,
  Dir = NULL,
  ASS = NULL,
  AssMod_names = NULL,
  compile = NULL,
  run = NULL,
  LastAssDat = NULL,
  ADMBpaths = NULL,
  make.comp = NULL,
  verbose = NULL
)
}
\arguments{
\item{Spc}{vector of strings specifying the name(s) of the stock
considered in the analysis.}

\item{GMACS_version}{vector of strings holding the name(s) of the
GMACS versions that is/are used in the analysis.}

\item{Dir}{vector of strings containing the directories for all
\code{GMACS_version} used in this analysis}

\item{ASS}{Logical. If TRUE, the outputs of the last assessment will be compared
to the GMACS version(s) currently used in this analysis.}

\item{AssMod_names}{vector of strings specifying the names of the models used
in the last assessment (e.g., model_16_0)}

\item{compile}{(0/1). If 0, GMACS is not compiled. This assumes that an
executable already exists in the directory of the version(s) used in the analysis.
If 1, the code will be compiled before a new run.}

\item{run}{Logical. If TRUE the model will be executed.}

\item{LastAssDat}{Logical. If TRUE, the latest available data will be used for
the analysis i.e. the model will be executed using the data stored in the
the "Assessment_data" folder.}

\item{ADMBpaths}{string name of 2-column text file that details the relevant
paths for the R variables admbpath, gccpath, and editor.}

\item{make.comp}{Logical. If TRUE, comparisons will be made between the various
\code{GMACS_version} considered in the analysis.}

\item{verbose}{logical: if \code{TRUE}, report the shell call an its messages to the R console.}
}
\description{
For each stock and version considered in the analysis, this function
realizes  the analysis based on the choice of the user, i.e: compiles the model,
runs the model, and/or establishes comparison between versions. This is the core
version to work with GMACS in `R`.
}
\examples{
\dontrun{
}

}
\seealso{
\code{\link{Do_GMACS}}, \code{\link{.buildGMACS}} for
building the executable.
}
