% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMACS_run_Utilities.R
\name{GMACSversion}
\alias{GMACSversion}
\title{Function to extract the Version number of GMACS}
\usage{
GMACSversion(DirTPL = NULL)
}
\arguments{
\item{DirTPL}{(character string)- the directory where the gmacsbase.TPL file
you are using for the stock assessment is hold.}
}
\value{
a named list with the variable \code{ver} that contains the version number
and \code{Comp} that holds the compilation date and time.
}
\description{
Extract the version number and the compilation date from the gmacs.tpl
}
