% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmr-package.R
\docType{package}
\name{gmr-package}
\alias{gmr-package}
\title{**gmr**: R code for GMACS}
\description{
Working with a Generalized size-structured Assessment Model for Crustaceans

The **gmr** package is a set of tools for running GMACS, a statistical size-structured 
assessment modeling framework for molting crustaceans species. The package 
allows analysing data and outputs related to GMACS stock assessment model
in the purpose of building stock assesmsent reports.
}
