% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertTime.R
\name{insertTime2}
\alias{insertTime2}
\title{insertTime2}
\usage{
insertTime2(object = NULL, pattern = NULL, .update = NULL)
}
\arguments{
\item{object}{Character string: Input vector/table}

\item{pattern}{Character string: pattern to look for.}

\item{.update}{Logical: is it to update a new version of GMACS? If TRUE,
the name of the new version of GMACS and the date of the compilation will be
modified in the gmacsbase.tpl file.}
}
\value{
List of object: the first object is the line wherein the text should
added (it corresponds to the line where the header is specified in gmacsbase.tpl).
The second object is the new header definition. If this is not an update of
GMACS, then the header of the gmacsbase.tpl file will not be modified but
the gmacs.tpl file will specify the date of the new compilation.
}
\description{
Function to insert time and date of the last compilation in the
gmacs.tpl file.
}
