% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewGMACSFeat.R
\name{NewGMACSFeat}
\alias{NewGMACSFeat}
\title{Write details of new implementations in the gmabase.TPL file}
\usage{
NewGMACSFeat(dirSrc, updateGMACS = NULL)
}
\arguments{
\item{updateGMACS}{(logical)- Flag specifying if the development version of GMACS
constitutes an update of the current version (in which case the version number
remains identical) or an upgrade version (in which case the version number
has to be incremented).}

\item{Dir}{(character string)- path of the folder name that holds the version
of GMACS that has been new developed.}
}
\value{
Add descriptive lines at the end of the gmacsbase.tpl file to detail
the new features that have been implemented.
}
\description{
This function ask the user to specify the details of the
new features implemented in the new version of GMACS.
}
