% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-meanSize.R, R/z_plot-catch.R
\name{plot_catch}
\alias{plot_catch}
\title{Plot observed and predicted catch values}
\usage{
plot_catch(
  M,
  plot_res = FALSE,
  scales = "free_y",
  xlab = "Year",
  ylab = "Catch",
  mlab = "Model",
  x_leg = 0.9,
  y_leg = 0.7
)

plot_catch(
  M,
  plot_res = FALSE,
  scales = "free_y",
  xlab = "Year",
  ylab = "Catch",
  mlab = "Model",
  x_leg = 0.9,
  y_leg = 0.7
)
}
\arguments{
\item{M}{list object created by read_admb function}

\item{plot_res}{plot residuals only (default = FALSE)}

\item{xlab}{the x-axis label for the plot}

\item{ylab}{the y-axis label for the plot}

\item{mlab}{the model label for the plot that appears above the key}
}
\value{
plot of catch history (observed) and predicted values

plot of catch history (observed) and predicted values
}
\description{
Plot observed and predicted catch values

Plot observed and predicted catch values
}
\author{
SJD Martell, D'Arcy N. Webber

SJD Martell, D'Arcy N. Webber
}
