% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_cpue_res}
\alias{plot_cpue_res}
\title{Plot residuals of cpue or other indices}
\usage{
plot_cpue_res(M)
}
\arguments{
\item{replist}{List object created by read_admb function}
}
\value{
Plot of fit indices residuals
}
\description{
Plot residuals of cpue or other indices
}
\author{
SJD Martell, DN Webber
}

