% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fishing-mortality.R
\name{plot_fishing_mortality}
\alias{plot_fishing_mortality}
\title{Plot fishing mortality (F)}
\usage{
plot_fishing_mortality(M)
}
\arguments{
\item{M}{list of object(s) created by read_admb function}
}
\value{
plot of fishing mortality (F)
}
\description{
TODO: complete this function!
}
\author{
JN Ianelli, SJD Martell, DN Webber
}
