% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gmr.R
\name{plot_gmr}
\alias{plot_gmr}
\title{Do all gmr plots}
\usage{
plot_gmr(M, plot_res = FALSE, target_dir = "",
  gmr_options = .gmr_options)
}
\arguments{
\item{M}{a list object created by read_admb function}

\item{plot_res}{plot residuals only (default = FALSE)}

\item{gmr_options}{the default plotting options}
}
\value{
All gmr plots.
}
\description{
Do all gmr plots
}
\author{
DN Webber
}
