% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_growth}
\alias{plot_growth}
\title{Plot growth from arbitrary start age}
\usage{
plot_growth(replist)
}
\arguments{
\item{replist}{List object created by read_admb function}
}
\value{
Plot natural mortality over time and size
}
\description{
Plot growth from arbitrary start age
}

