% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-growth-inc.R, R/plot-growth-tag.R
\name{plot_growth_inc}
\alias{plot_growth_inc}
\title{Plot growth from arbitrary start age}
\usage{
plot_growth_inc(
  M,
  xlab = "Pre-molt size (mm)",
  ylab = "Molt increment (mm)",
  slab = "Sex"
)

plot_growth_inc(
  M,
  xlab = "Pre-molt size (mm)",
  ylab = "Molt increment (mm)",
  slab = "Sex"
)
}
\arguments{
\item{M}{list object created by read_admb function}

\item{xlab}{the x-axis label for the plot}

\item{ylab}{the y-axis label for the plot}

\item{slab}{the sex label for the plot that appears above the key}
}
\value{
plot growth increment for given pre-molt size, including model predictions and data

plot growth increment for given pre-molt size, including model predictions and data
}
\description{
Plot growth from arbitrary start age

Plot growth from arbitrary start age
}
\author{
SJD Martell, D'Arcy N. Webber

SJD Martell, D'Arcy N. Webber
}
