% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-growth-ridges.R
\name{plot_growth_ridges}
\alias{plot_growth_ridges}
\title{Plot growth transition matrices as ridges}
\usage{
plot_growth_ridges(M, showPlot = FALSE)
}
\arguments{
\item{M}{- list with model results objects from [read_admb()] as named elements}
}
\value{
ggplot2 object using [ggridges::geom_density_ridges()] to visualize
p(post-molt size|pre-molt size).
}
\description{
Function to plot growth transition matrices as ridges.
}
\author{
WT Stockhausen (ctb)
}
