% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-size-transition.R
\name{plot_growth_transition}
\alias{plot_growth_transition}
\title{Plot growth transition}
\usage{
plot_growth_transition(M)
}
\arguments{
\item{M}{list of object(s) created by read_admb function}
}
\value{
plot of size transition matrix
}
\description{
Plot growth transition
}
\author{
SJD Martell, DN Webber
}

