% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-kobe.R
\name{plot_kobe}
\alias{plot_kobe}
\title{Plot predicted spawning stock biomass (ssb)}
\usage{
plot_kobe(
  M,
  ylab = "Fishing mortality",
  xlab = "SSB (tonnes)",
  ylim = NULL,
  alpha = 0.1,
  sex = "Male",
  fleet_in = "Pot",
  ref_ind = 1
)
}
\arguments{
\item{M}{List object(s) created by read_admb function}

\item{ylab}{the y-label of the figure}

\item{xlab}{the x-label of the figure}

\item{ylim}{is the upper limit of the figure}

\item{alpha}{the opacity of the ribbon}
}
\value{
Plot of model estimates of spawning stock biomass
}
\description{
Spawning biomass may be defined as all males or some combination of males and
females
}
