% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-maturity.R
\name{plot_length_maturity}
\alias{plot_length_maturity}
\title{Plot length-maturity relationship}
\usage{
plot_length_maturity(M, xlab = "Size", ylab = "Proportion mature")
}
\arguments{
\item{M}{list object created by read_admb function}

\item{xlab}{the x-axis label for the plot}

\item{ylab}{the y-axis label for the plot}
}
\value{
plot of the length-weight relationship
}
\description{
Plot length-maturity relationship
}
\author{
D'Arcy N. Webber
}
