% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-length-weight.R
\name{plot_length_weight}
\alias{plot_length_weight}
\title{Plot length-weight relationship}
\usage{
plot_length_weight(M)
}
\arguments{
\item{M}{list object created by read_admb function}
}
\value{
plot of the length-weight relationship
}
\description{
Plot length-weight relationship
}
\author{
DN Webber
}

