% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-recruitment.R
\name{plot_models_recruitment}
\alias{plot_models_recruitment}
\title{Plot predicted recruitment across model runs}
\usage{
plot_models_recruitment(data, modnames = NULL)
}
\arguments{
\item{data}{A list of multiple objects created by read_admb function}

\item{modnames}{A vector of model names included in \code{data}}
}
\value{
Plot of predicted recruitment compared across models
}
\description{
Plot predicted recruitment across model runs
}
\author{
Cole Monnahan Kelli Johnson
}

