% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-molt-prob.R
\name{plot_molt_prob}
\alias{plot_molt_prob}
\title{Plot molting probability relationship}
\usage{
plot_molt_prob(M, xlab = "Length", ylab = "Probability")
}
\arguments{
\item{M}{list object created by read_admb function}

\item{xlab}{the x-axis label for the plot}

\item{ylab}{the y-axis label for the plot}
}
\value{
plot of the molting probability relationship
}
\description{
Plot molting probability relationship
}
\author{
D'Arcy N. Webber
}

