% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_natural_mortality}
\alias{plot_natural_mortality}
\title{Plot natural mortality}
\usage{
plot_natural_mortality(M, plt_surface = FALSE)
}
\arguments{
\item{M}{list object created by read_admb function}

\item{plt_surface}{include a panel with surface over size-time}
}
\value{
Plot natural mortality over time (and size)
}
\description{
Plot natural mortality
}
\author{
J Ianelli, SJD Martell, DN Webber
}

