% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_numbers}
\alias{plot_numbers}
\title{Plot numbers at length}
\usage{
plot_numbers(M, subsetby = "")
}
\arguments{
\item{M}{list object created by read_admb function}

\item{subsetby}{a selection of the years to restrict plotting of}
}
\value{
plot of recruitment size distribution
}
\description{
Plot numbers at length
}
\author{
DN Webber
}

