% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-recruitment.R
\name{plot_recruitment_size}
\alias{plot_recruitment_size}
\title{Plot recruitment size distribution}
\usage{
plot_recruitment_size(
  M,
  xlab = "Mid-point of size class (mm)",
  ylab = "Proportion recruiting"
)
}
\arguments{
\item{M}{list object created by read_admb function}

\item{xlab}{the x-axis label for the plot}

\item{ylab}{the y-axis label for the plot}
}
\value{
recruitment size distribution plot as ggplot2 object
}
\description{
Plot recruitment size distribution
}
\author{
DN Webber, WT Stockhausen [ctb]
}
