% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-selectivity.R
\name{plot_selectivity}
\alias{plot_selectivity}
\title{Plot selectivity}
\usage{
plot_selectivity(M, xlab = "Mid-point of size class (mm)",
  ylab = "Selectivity", tlab = "Type", ilab = "Period year",
  nrow = NULL, ncol = NULL)
}
\arguments{
\item{M}{list of lists created by the read_admb function}

\item{xlab}{the x-axis label for the plot}

\item{ylab}{the y-axis label for the plot}

\item{tlab}{the type (retained or discarded) label for the plot that appears above the key}

\item{ilab}{the year label for the plot that appears above the key}

\item{nrow}{the number of rows in the facet grid}

\item{ncol}{the number of columns in the facet grid}
}
\value{
plot of selectivity
}
\description{
This function takes a list of lists created by the read_admb function plots
the selectivity by fishing fleet, sex, year, type (retained or discarded) and
model.
}
\author{
SJD Martell, D'Arcy N. Webber
}
