% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-selectivity.R
\name{plot_selectivity_3d}
\alias{plot_selectivity_3d}
\title{Plot selectivity 3D}
\usage{
plot_selectivity_3d(M, plt_surface = FALSE)
}
\arguments{
\item{M}{list object created by read_admb function}

\item{plt_surface}{include a panel with surface over size-time}
}
\value{
3D plot of selectivity
}
\description{
Plot selectivity 3D
}
\author{
D'Arcy N. Webber
}
