% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_size_comps}
\alias{plot_size_comps}
\title{plot_sizeComps}
\usage{
plot_size_comps(M, which_plots = "all")
}
\arguments{
\item{M}{List object(s) created by read_admb function}
}
\value{
Plots of observed and predicted size composition values
}
\description{
Get observed and predicted size composition values
}
\author{
SJD Martell
}

