% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-size-comps.R
\name{plot_size_comps_res}
\alias{plot_size_comps_res}
\title{Plot fits to size composition data}
\usage{
plot_size_comps_res(M, ncol = 1, xlab = "Year",
  ylab = "Mid-point of size-class (mm)")
}
\arguments{
\item{M}{List object(s) created by read_admb function}

\item{xlab}{the x-axis label for the plot}

\item{ylab}{the y-axis label for the plot}
}
\value{
Plots of observed and predicted size composition values
}
\description{
Get observed and predicted size composition values
}
