% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-size-comps.R
\name{plot_size_ridges}
\alias{plot_size_ridges}
\title{Plot fits to size composition data}
\usage{
plot_size_ridges(
  M,
  which_plots = "all",
  xlab = "Mid-point of size-class (mm)",
  ylab = "Proportion",
  slab = "Sex",
  mlab = "Model",
  tlab = "Fleet",
  res = FALSE,
  legend_loc = c(1, 1),
  ylim_max = 0.3
)
}
\arguments{
\item{M}{List object(s) created by read_admb function}

\item{which_plots}{the size composition fits that you want to plot}

\item{xlab}{the x-axis label for the plot}

\item{ylab}{the y-axis label for the plot}

\item{slab}{the sex label for the plot that appears above the key}

\item{mlab}{the model label for the plot that appears above the key}

\item{tlab}{the fleet label for the plot that appears above the key}

\item{res}{boolean if residual or observed and predicted}

\item{legend_loc}{(numeric vector); c(x,y) where x and y are the coordinates of the legend box}

\item{ylim_max}{(numeric), defines the ylim.}
}
\value{
Plots of observed and predicted size composition values
}
\description{
Get observed and predicted size composition values
}
