% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-transitions.R
\name{plot_size_transition}
\alias{plot_size_transition}
\title{Plot size transition}
\usage{
plot_size_transition(M, xlab = "Size-class",
  ylab = "P(size transition | molt)")
}
\arguments{
\item{M}{list of object(s) created by read_admb function}

\item{xlab}{the x-axis label for the plot}

\item{ylab}{the y-axis label for the plot}
}
\value{
plot of size transition matrix
}
\description{
The combination of growth transitions and molting probability represents the
size transition.
}
\author{
SJD Martell, DN Webber
}

