% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-size-comps.R
\name{plot_sizeCompRes}
\alias{plot_sizeCompRes}
\title{plot_sizeCompRes
Get observed and predicted size composition values}
\usage{
plot_sizeCompRes(M, which.plot = "all")
}
\arguments{
\item{M}{List object(s) created by read_admb function}
}
\value{
Plots of observed and predicted size composition values
}
\description{
plot_sizeCompRes
Get observed and predicted size composition values
}
\author{
SJD Martell
}

