% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-sizecomp.R
\name{plot_sizeComps}
\alias{plot_sizeComps}
\title{plot_sizeComps
Get observed and predicted size composition values}
\usage{
plot_sizeComps(M, which_plots = "all")
}
\arguments{
\item{M}{List object(s) created by read_admb function}
}
\value{
Plots of observed and predicted size composition values
}
\description{
plot_sizeComps
Get observed and predicted size composition values
}
\author{
SJD Martell
}

