% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-sizetransition.R
\name{plot_sizeTransition}
\alias{plot_sizeTransition}
\title{Plot size transition}
\usage{
plot_sizeTransition(M)
}
\arguments{
\item{M}{List of object(s) created by read_admb function}
}
\value{
Plot of size transition matrix
}
\description{
Plot size transition
}
\author{
SJD Martell
}

