% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-type.R
\name{plot_type}
\alias{plot_type}
\title{Specify the type of plot you would like to create}
\usage{
plot_type(filename, width, height, gmr_options = .gmr_options)
}
\arguments{
\item{filename}{the directory and name of the file to save}

\item{width}{the width of the plot in mm}

\item{height}{the height of the plot in mm}

\item{gmr_options}{a list of options for plotting}
}
\value{
a plot device
}
\description{
Specify the type of plot you would like to create
}
\author{
DN Webber
}

