% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_module_functions.R
\name{prepSim}
\alias{prepSim}
\title{Get the original input files for the simulation approach}
\usage{
prepSim(Dir = NULL, Spc = NULL, verbose = NULL)
}
\arguments{
\item{Dir}{(character string)- path to the repertory that holds the \code{"build"} folder.}

\item{Spc}{(character string)- the name of the stock of interest for the simulations}

\item{verbose}{(logical)- flag to print processing information}
}
\value{
A named list with the following arguments:
\itemize{
\item \code{Dirlist}: named list of the working directory used and/or created,
\item \code{GmacsFile}: the gmacs.dat file as a named list,
\item \code{DatFile}: the Spc.dat file as a named list,
\item \code{CtlFile}: the Spc.ctl file as a named list,
\item \code{PrjFile}: the Spc.pj file as a named list,
\item \code{ParFile}: the gmacs.par file as a named list,
\item \code{OrigSimData}: the simdata.out file (the original one) as a named list.
}
}
\description{
Prepares the simulations:
\enumerate{
\item Creates a specific folder (\code{Simulations}) in the designated repertory of
the stock of interest
\item Creates two sub folders (\code{True} and \code{Runs}) in the
\code{Simulations} folder to store the outputs of the simulations
\item Reads the original input files that will be modified to simulate the data
}
}
\seealso{
\code{\link{readGMACSfiles}}, \code{\link{readGMACSdat}},
\code{\link{readGMACSctl}}, \code{\link{readGMACSprj}}, \code{\link{readGMACSpar}},
\code{\link{readGMACSsimdat}}
}
