% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-OUT.R
\name{read.OUT}
\alias{read.OUT}
\title{read.OUT}
\usage{
read.OUT(file)
}
\arguments{
\item{file}{Name (with directory) of the gmacsall.out}
}
\value{
list of management/biological quantities
}
\description{
This function allows to read the gmacsall.OUT to find estimates
of specific management/biological quantities

Need to be modified to allow a vector of character to search any variable
}
