% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-admb.R
\name{read_admb}
\alias{read_admb}
\title{Read ADMB output files}
\usage{
read_admb(repfile)
}
\arguments{
\item{repfile}{ADMB output files to be read (no extension needed)}
}
\value{
object of type 'list' with ADMB outputs as list elements
}
\description{
Read ADMB output files .rep, .par, and .cor and return an R object of type 'list'
}
\author{
Steve Martell, Anders Nielsen, Athol Whitten, D'Arcy N. Webber
}
