% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-admb.R
\name{read_ctl}
\alias{read_ctl}
\title{Read ADMB .ctl file}
\usage{
read_ctl(fn)
}
\arguments{
\item{repfile}{name of ADMB output file to be read}
}
\value{
object of type 'list' with ADMB outputs therein
}
\description{
Read ADMB .ctl file and return an R object of type 'list'. DOES NOT WORK
}
\author{
D'Arcy N. Webber
}
