% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-admb.R
\name{read_fit}
\alias{read_fit}
\title{Read ADMB .par, .std, and .cor file and return an R object of type 'list' of estimates and correlations}
\usage{
read_fit(repfile)
}
\arguments{
\item{repfile}{name of ADMB output file to be read (no extension needed)}
}
\value{
object of type 'list' with ADMB outputs therein
}
\description{
Read ADMB .par, .std, and .cor file and return an R object of type 'list' of estimates and correlations
}
\author{
Steve Martell, Anders Nielsen, Athol Whitten, D'Arcy N. Webber
}
