% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-admb.R
\name{read_psv}
\alias{read_psv}
\title{Read ADMB .psv file}
\usage{
read_psv(fn, nsamples = 10000)
}
\arguments{
\item{repfile}{name of ADMB output file to be read (no extension needed)}
}
\value{
object of type 'list' with ADMB outputs therein
}
\description{
Read ADMB .psv file and return an R object of type 'list'
}
\author{
Steve Martell
}

