% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read-admb.R
\name{read_rep}
\alias{read_rep}
\title{Read ADMB .rep file}
\usage{
read_rep(fn)
}
\arguments{
\item{fn}{name of ADMB output file to be read (no extension needed)}
}
\value{
object of type "list" with ADMB outputs therein
}
\description{
Read ADMB .rep file and return an R object of type 'list'
}
\author{
Steve Martell, D'Arcy N. Webber
}

