% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGMACSallOUT.R
\name{readGMACSallOUT}
\alias{readGMACSallOUT}
\title{Read GmacsAll.out}
\usage{
readGMACSallOUT(
  FileName = NULL,
  verbose = TRUE,
  DatFile = NULL,
  CtlFile = NULL,
  GmacsFile = NULL,
  nyrRetro = NULL
)
}
\arguments{
\item{FileName}{(character string)- path and name (by default, gmacsAll.out) of the file to read.}

\item{verbose}{(logical)- (TRUE/FALSE); flag to print processing information}

\item{DatFile}{(list)- Object containing the .dat file - This is the output
of the \code{\link[=readGMACSdat]{readGMACSdat()}} function.}

\item{CtlFile}{(list)- Object containing the .ctl file - This is the output
of the \code{\link[=readGMACSctl]{readGMACSctl()}} function.}

\item{GmacsFile}{(list)- Object containing the gmacs.dat file - This is the output
of the \code{\link[=readGMACS.dat]{readGMACS.dat()}} function.}

\item{nyrRetro}{(integer)- Number of year for the retrospective analysis}
}
\value{
the gmacsall.out file as a named list.
}
\description{
This functions reads the gmacsall.out file as a named list
within R.
}
\seealso{
\code{\link{readGMACSdat}}, \code{\link{readGMACSctl}}, \code{\link{readGMACSrep}}
}
