% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGMACSfiles.R
\name{readGMACSfiles}
\alias{readGMACSfiles}
\title{Read all the GMACS input files}
\usage{
readGMACSfiles(Dir = NULL, verbose = NULL)
}
\arguments{
\item{verbose}{\itemize{
\item (TRUE/FALSE); flag to print processing information
}}

\item{FileName}{\itemize{
\item path to the folder containing the GMACS input files
}}
}
\value{
the gmacs.dat; .dat, .ctl, and .prj files, as a named list.
}
\description{
Function that read all the GMACS input files.
}
\details{
the following names are used in the list returned by this function
\itemize{
\item GmacsFile: list holding the gmacs.dat file
\item DatFile: list holding the Spc.dat file
\item CtlFile: list holding the Spc.ctl file
\item PrjFile: list holding the Spc.prj file
where Spc designs the stock of interest.
}
}
\seealso{
\code{\link{readGMACS.dat}},\code{\link{readGMACSdat}},
\code{\link{readGMACSctl}},\code{\link{readGMACSprj}}
}
