% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGMACSprj.R
\name{readGMACSprj}
\alias{readGMACSprj}
\title{readGMACSprj}
\usage{
readGMACSprj(FileName = NULL, verbose = TRUE)
}
\arguments{
\item{FileName}{\itemize{
\item path (and name) of the projection file (e.g. snow.prj)
}}

\item{verbose}{\itemize{
\item (TRUE/FALSE); flag to print processing information
}}
}
\value{
the .prj file as a named list.
}
\description{
Read the GMACS projection file.
}
\seealso{
\code{\link{readGMACS.dat}},\code{\link{readGMACSdat}},
\code{\link{readGMACSctl}}
}
