% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGMACSsimdat.R
\name{readGMACSsimdat}
\alias{readGMACSsimdat}
\title{readGMACSsimdat}
\usage{
readGMACSsimdat(
  Dir = NULL,
  FileName = "simdata.out",
  verbose = TRUE,
  DatFile = NULL,
  CtlFile = NULL
)
}
\arguments{
\item{Dir}{\itemize{
\item path to the "simulated" data file (output from GMACS)
}}

\item{FileName}{\itemize{
\item name of the "simulated" data file - By default, "simdata.out"
}}

\item{verbose}{\itemize{
\item (TRUE/FALSE); flag to print processing information
}}

\item{DatFile}{\itemize{
\item Object (list) containing the .dat file - This is the output
of the \code{\link[=readGMACSdat]{readGMACSdat()}} function.
}}

\item{CtlFile}{\itemize{
\item Object (list) containing the .ctl file - This is the output
of the \code{\link[=readGMACSctl]{readGMACSctl()}} function.
}}
}
\value{
the simdata.out file as a named list.
}
\description{
Read the simdata.out file. This is an output of GMACS that holds
the predicted catch and cpue from an assessment run as well as the observed
and predicted size compositions.
}
\seealso{
\code{\link{readGMACS.dat}},\code{\link{readGMACSdat}},\code{\link{readGMACSctl}},
\code{\link{readGMACSprj}}, \code{\link{readGMACSpar}}
}
