% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_module_functions.R
\name{RunGmacsSim}
\alias{RunGmacsSim}
\title{Run the GMACS simulation-estimation}
\usage{
RunGmacsSim(path = NULL, Stock = NULL, Nsim = NULL, verbose = NULL)
}
\arguments{
\item{path}{(character string)- file path representing root i.e., the folder where
the \code{gmacsbase.tpl} file is stored.}

\item{Stock}{(character string)- Name of the stock of interest. This will be used
to set the working directories: this is the name of the folder in the
\code{build} directory where the basic gmacs input files are stored.}

\item{Nsim}{(integer)- the number of simulation that have been considered in
the analysis (see \code{\link[=Gen_GmacsSim]{Gen_GmacsSim()}}).}

\item{verbose}{(logical)- flag to print processing information}
}
\value{
Nothing
}
\description{
Executes the simulations with Gmacs using the data generated with
the \code{\link[=Gen_GmacsSim]{Gen_GmacsSim()}} function.
}
