% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_module_functions.R
\name{SaveSimFiles}
\alias{SaveSimFiles}
\title{Save the outputs of each simulation-estimation run}
\usage{
SaveSimFiles(
  simDat = NULL,
  UsePar = NULL,
  NoEst = NULL,
  Isim = NULL,
  Nsim = NULL,
  Dir_TPL = NULL
)
}
\arguments{
\item{simDat}{(list)- the output of the \code{\link[=SimData]{SimData()}} function for each simulation.}

\item{UsePar}{(logical)- Do the simulations use a .PIN file to import parameters?}

\item{NoEst}{(logical)- Set the maximum function call to 1? i.e., run estimation?}

\item{Isim}{(integer)- the current simulation.}

\item{Nsim}{(integer)- the number of simulation to run.}

\item{Dir_TPL}{(character string)- the directory where the gmacsbase.TPL file
you are using for the stock assessment is hold.}
}
\value{
Nothing
}
\description{
Save the outputs of each simulation

For each simulation the function writes the gmacs.dat, data, control, and
projection files in the \code{Runs} folder created by the \code{\link[=prepSim]{prepSim()}} functions.
The function also writes a batch command that will be called to when running
Gmacs (name of the file \code{RunSim.bat}).
}
