% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set-ggtheme.R
\name{set_ggtheme}
\alias{set_ggtheme}
\title{Set plotting theme for ggplot2 via gmr}
\usage{
set_ggtheme(theme)
}
\arguments{
\item{name}{of desired theme}
}
\value{
Sets ggplot2 theme for current working session
}
\description{
Gives user control over plot theme by running ggplot2 functions
that do the same. This allows a user to set the theme without
independently loading the ggplot2 package.
}

