% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetGMACSfold.R
\name{SetGMACSfold}
\alias{SetGMACSfold}
\title{Create the needed folders to run GMACS}
\usage{
SetGMACSfold(vv = NULL, Dir = NULL, Spc = NULL, Sp = NULL, verbose = FALSE)
}
\arguments{
\item{vv}{\itemize{
\item integer index in \code{Dir} indicating root folder to create these
base folders.
}}

\item{Dir}{\itemize{
\item vector of root folder names
}}

\item{Spc}{\itemize{
\item vector of species names
}}

\item{Sp}{\itemize{
\item integer index in \code{Spc} indicating the name(s) of the species
for which the folder needs to be created. If \code{Sp = "all"}, then a
folder will be created for all species designated in \code{Spc}.
}}

\item{verbose}{\itemize{
\item flag (TRUE/FALSE) to print processing information
}}
}
\value{
nothing
}
\description{
Function to create empty folders needed to run GMACS: \code{/build},
\code{/build/debug}, \code{/build/release}, and a folder for the stock(S)
(\code{Spc}) of interest (\code{/build/Spc}).
}
\details{
Both the input and output files of GMACS will be contained in the
folder whose name corresponds to the species of interest. When starting from
scratch, you will need to provide this folder with all the assessment files
needed by GMACS (.dat, .ctl, and .prj files).
}
