% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_gmacs.R
\name{shiny_gmacs}
\alias{shiny_gmacs}
\title{Plot Gmacs on shiny app}
\usage{
shiny_gmacs(gmrep)
}
\arguments{
\item{replist}{List object created by read_admb function}
}
\description{
Plot Gmacs on shiny app
}

