% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_module_functions.R
\name{SimData}
\alias{SimData}
\title{Generate the simulated data}
\usage{
SimData(Outres = NULL, Rand = NULL)
}
\arguments{
\item{Outres}{(named list)- list that holds the outputs of the \code{\link[=prepSim]{prepSim()}} and
\code{\link[=GenSimFiles]{GenSimFiles()}} functions.}

\item{Rand}{(logical)- If true, random deviations are generated.}
}
\value{
a list that hold the simulated data.
}
\description{
Generates the data for the simulation. The functions generate:
\enumerate{
\item catch data,
\item Index/cpue data,
\item length frequency data.
}

All these data can either be considered as deterministic or random generation.
When noise is added, the function generates normally distributed deviations for
catch and cpue data while the deviations for the size composition data
can be generated either from a a multinormal or a dirichlet distribution.
The distribution considered is based on the specifications in the control file.
}
