% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-catch.R
\name{table_catch}
\alias{table_catch}
\title{Create a table of catches}
\usage{
table_catch(M)
}
\arguments{
\item{M}{List object(s) created by read_admb function}
}
\value{
kable of catches per year and fleet
}
\description{
Create a table of catches
}
\author{
SJD Martell, DN Webber
}
