% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/table-likelihoods.R
\name{table_likelihoods}
\alias{table_likelihoods}
\title{Create a table of likelihoods}
\usage{
table_likelihoods(M)
}
\arguments{
\item{M}{list object(s) created by read_admb function}
}
\value{
dataframe of likelihood components
}
\description{
Create a table of likelihoods
}
\author{
D'Arcy N. Webber
}

