% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/table-likelihoods.R
\name{table_priors}
\alias{table_priors}
\title{Create a table of priors}
\usage{
table_priors(M)
}
\arguments{
\item{M}{list object(s) created by read_admb function}
}
\value{
dataframe of prior components
}
\description{
Create a table of priors
}
\author{
D'Arcy N. Webber
}

