% Generated by roxygen2 (4.0.2): do not edit by hand
\name{table_spr}
\alias{table_spr}
\title{Create a table of SPR reference points}
\usage{
table_spr(M)
}
\arguments{
\item{M}{List object(s) created by read_admb function}
}
\value{
dataframe of SPR reference points
}
\description{
Create a table of SPR reference points
}
\author{
SJD Martell, DN Webber
}

