% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateGMACS.R
\name{UpdateGMACS}
\alias{UpdateGMACS}
\title{UpdateGMACS}
\usage{
UpdateGMACS()
}
\description{
Function used when updating and releasing a new version of GMACS.
It allows to copy and paste all the GMACS files (for one or several stocks)
resulting from the Development version from this directory to the Latest_Version
directory.
}
