% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeGmacs_dat.R
\name{writeGmacs.dat}
\alias{writeGmacs.dat}
\title{Write the gmacs.dat file}
\usage{
writeGmacs.dat(
  Dir = NULL,
  FileName = NULL,
  gmacsDat = NULL,
  stock = "",
  model_name = "",
  Ass_Year = "",
  DirTPL = NULL,
  Ver_number = NULL
)
}
\arguments{
\item{Dir}{(character string)- directory where to save the new gmacs.dat file.}

\item{FileName}{(character string)- name of the new gmacs.dat file.}

\item{gmacsDat}{(character string)- Object (list) containing the ex gmacs.dat
file - The list is created using the \code{\link[=readGMACS_dat]{readGMACS_dat()}} function.}

\item{stock}{(character string)- name of the stock of interest.}

\item{model_name}{(character string)- name of the model currently considered (e.g., "model 22.A").}

\item{Ass_Year}{(character string)- Year of this assessment.}

\item{DirTPL}{(character string)- the directory where the gmacsbase.TPL file
you are using for the stock assessment is hold.}

\item{Ver_number}{(character)- The version of Gmacs. This is used only when
developing new code versions of Gmacs.}
}
\value{
create a new gmacs.dat file.
}
\description{
Write a new gmacs.dat file. This function is used to modify within
R a pre-existent gmacs.dat file.
}
\seealso{
\code{\link{readGMACS.dat}}
}
