% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeGmacsctlfile.R
\name{writeGmacsctlfile}
\alias{writeGmacsctlfile}
\title{Write the control file}
\usage{
writeGmacsctlfile(
  Dir = NULL,
  FileName = NULL,
  CtlFile = NULL,
  DatFile = NULL,
  stock = "",
  model_name = "",
  Ass_Year = "",
  nsex = NULL,
  nmature = NULL,
  Start_Y = NULL,
  End_Y = NULL,
  DirTPL = NULL
)
}
\arguments{
\item{Dir}{(character string)- path where to save the new Spc.ctl file}

\item{FileName}{(character string)- name of the new Spc.ctl file}

\item{CtlFile}{(character string)- Object (list) containing the ex Spc.ctl file - The list is
created using the \code{\link[=readGMACSctl]{readGMACSctl()}} function.}

\item{DatFile}{(list; optional)- Object containing the .dat file - This can either be the output
of the \code{\link[=readGMACSdat]{readGMACSdat()}} function. If \code{DatFile} is provided, then you don't need to provide
the following: \code{nsex; Start_Y; End_Y}}

\item{stock}{(character string)- name of the stock of interest}

\item{model_name}{(character string)- name of the model currently considered (e.g., "model 22.A")}

\item{Ass_Year}{(character string)- Year of this assessment}

\item{nsex}{(integer; optional)- Number of sexes considered in the model. This is required if you
don't provide the \code{DatFile}.}

\item{nmature}{(integer; optional)- Number of maturity type(s) considered in the model.
This is required if you don't provide the \code{DatFile}.}

\item{DirTPL}{(character string)- the directory where the gmacsbase.TPL file
you are using for the stock assessment is hold.}

\item{Start_Y;End_Y}{(integer; optional)- First and last year of the assessment period.
This is required if you don't provide the \code{DatFile}.}
}
\value{
create a new .ctl file.
}
\description{
Write a new Spc.ctl file. This function is used to modify (within
R) a pre-existent Spc.ctl file.
}
\seealso{
\code{\link{readGMACSctl}}
}
