% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_files.R
\name{clean_bat}
\alias{clean_bat}
\title{Clean gmacs output files in a folder}
\usage{
clean_bat(path = ".", verbose = TRUE)
}
\arguments{
\item{path}{character string or file path representing root (default: ".")}

\item{verbose}{(TRUE/FALSE) flag to print file names being deleted}
}
\value{
nothing
}
\description{
Function to clean gmacs output files in a folder specified by \code{path}.
By default, this is the current working directory.
}
\details{

}
