% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_files.R
\name{clean_bat_Sim}
\alias{clean_bat_Sim}
\title{Clean the output files of the Gmacs simulation module}
\usage{
clean_bat_Sim(path = ".", verbose = TRUE)
}
\arguments{
\item{path}{(character string)- names or file path representing root (default: ".")}

\item{verbose}{(TRUE/FALSE) flag to print file names being deleted}
}
\value{
nothing
}
\description{
Function to clean gmacs output files in a folder when a
simulation-estimation approach has been considered.
}
\details{

}
