% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_files.R
\name{clean_files}
\alias{clean_files}
\title{Clean files in a folder}
\usage{
clean_files(path = ".", filename = NULL, names = c("no_file"), verbose = FALSE)
}
\arguments{
\item{path}{(path)- path to the root folder to be cleaned.}

\item{filename}{(character vector)- file name(s) to be removed.}

\item{names}{(character vector)- character strings representing the files in
the root folder to be removed.}

\item{verbose}{(logical)- flag to print file names being deleted}
}
\value{
nothing
}
\description{
Function to clean files in a folder.
}
\details{
removes (deletes) files with given names on given path.
}
