% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGmacsExe.R
\name{.GetGmacsExe}
\alias{.GetGmacsExe}
\title{.GetGmacsExe}
\usage{
.GetGmacsExe()
}
\value{
the new GMACS executable
}
\description{
Function used when updating and releasing a new version of GMACS.
It uses the (\code{\link{write_TPL}}) function to editate the new
gmacs.tpl, compiles the model and uses the (\code{\link{.buildGMACS}})
function to provide a new executable.
}
\examples{
\dontrun{
}

}
