% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMACS_run_Utilities.R
\name{GMACSversion}
\alias{GMACSversion}
\title{Function to extract the Version number of GMACS}
\usage{
GMACSversion(Dir = NULL)
}
\arguments{
\item{Dir}{Character string: the directory where the gmacs.exe is hold for the stock
your are currently working on.}
}
\value{
a named list with the variable \code{ver} that contains the version number
and \code{Comp} that holds the compilation date and time.
}
\description{
Extract the version number and the compilation date from the gmacs.tpl
}
