% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basicOutput.R
\name{plot_basicOutput}
\alias{plot_basicOutput}
\title{Plotting the basic outputs of GMACS}
\usage{
plot_basicOutput(
  Stock = NULL,
  Dir = NULL,
  model_name = NULL,
  GMACSFileName = "gmacs.dat",
  out = TRUE,
  Dir.out = NULL,
  save.out = NULL,
  out.format = NULL
)
}
\arguments{
\item{Stock}{vector of strings specifying the name(s) of the stock
considered in the analysis.}

\item{Dir}{vector of strings containing the directories for all
\code{GMACS_version} used in this analysis.}

\item{model_name}{vector of strings holding the name(s) of the
GMACS versions/models that is/are used in the analysis.}

\item{GMACSFileName}{name of the \code{gmacs.dat} file.
Default \code{"gmacs.dat"}.}

\item{out}{(logical); if \code{TRUE}; the plot will be saved on the \code{Dir.out}
directory.}

\item{Dir.out}{(character string); path to the directory to save plots}

\item{save.out}{(character string);}

\item{out.format}{Device to use to save the plot (currently only "pdf")}
}
\description{
Plots the basic outputs of GMACS.
}
