% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_catch}
\alias{plot_catch}
\title{Plot observed and predicted catch values}
\usage{
plot_catch(M, plot_res = FALSE)
}
\arguments{
\item{M}{list object created by read_admb function}

\item{plot_res}{plot residuals only (default = FALSE)}
}
\value{
plot of catch history (observed) and predicted values
}
\description{
Plot observed and predicted catch values
}
\author{
SJD Martell, DN Webber
}

