% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-catch.R
\name{plot_catch}
\alias{plot_catch}
\title{Plot observed and predicted catch values}
\usage{
plot_catch(M, plot_res = FALSE)
}
\arguments{
\item{plot_res}{plot residuals only (default=F)}

\item{replist}{List object created by read_admb function}
}
\value{
Plot of catch history (observed) and predicted values
}
\description{
Plot observed and predicted catch values
}
\author{
SJD Martell, DN Webber
}

