% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-cpue.R
\name{plot_cpue}
\alias{plot_cpue}
\title{Plot cpue or other indices}
\usage{
plot_cpue(M, subsetby = "")
}
\arguments{
\item{M}{list object created by read_admb function}
}
\value{
plot of all observed and predicted incices
}
\description{
Plot cpue or other indices
}
\author{
SJD Martell, DN Webber
}

