% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cpue.R
\name{plot_cpue_res}
\alias{plot_cpue_res}
\title{Plot residuals of cpue or other indices}
\usage{
plot_cpue_res(M, xlab = "Year", ylab = "Residual (Observed - Predicted)",
  slab = "Sex")
}
\arguments{
\item{xlab}{the x-axis label for the plot}

\item{ylab}{the y-axis label for the plot}

\item{slab}{the sex label for the plot that appears above the key}

\item{replist}{List object created by read_admb function}
}
\value{
plot of fit indices residuals
}
\description{
Plot residuals of cpue or other indices
}
\author{
SJD Martell, D'Arcy N. Webber
}

