% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_datarange}
\alias{plot_datarange}
\title{Plot data range by fleet and year}
\usage{
plot_datarange(M)
}
\arguments{
\item{M}{List object(s) created by read_admb function}
}
\value{
Plot of data range
}
\description{
Plot data range by fleet and year
}
\author{
Ian Taylor, Huihua Lee, Jim Ianelli
}

