% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-datarange.R
\name{plot_datarange}
\alias{plot_datarange}
\title{Plot data range by fleet and year}
\usage{
plot_datarange(M)
}
\arguments{
\item{M}{a list of lists object created by the read_admb function}
}
\value{
plot of data range
}
\description{
Plot data range by fleet and year
}
\author{
Ian Taylor, Huihua Lee, Jim Ianelli, D'Arcy N. Webber
}

