% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-F.R
\name{plot_F}
\alias{plot_F}
\title{Plot fishing mortality}
\usage{
plot_F(M, scales = "free_y", xlab = "Year", ylab = "F", mlab = "Model")
}
\arguments{
\item{M}{list object created by read_admb function}

\item{xlab}{the x-axis label for the plot}

\item{ylab}{the y-axis label for the plot}

\item{mlab}{the model label for the plot that appears above the key}
}
\value{
plot of fishing mortality and mean Fs
}
\description{
Plot fishing mortality
}
\author{
D'Arcy N. Webber
}

