% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_fishing_mortality}
\alias{plot_fishing_mortality}
\title{Plot fishing mortality (F)}
\usage{
plot_fishing_mortality(M)
}
\arguments{
\item{M}{list of object(s) created by read_admb function}

\item{xlab}{the x-axis label for the plot}

\item{ylab}{the y-axis label for the plot}
}
\value{
plot of fishing mortality (F)
}
\description{
Plot fishing mortality (F)
}
\author{
SJD Martell, DN Webber
}

