% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_growth_inc}
\alias{plot_growth_inc}
\title{Plot growth from arbitrary start age}
\usage{
plot_growth_inc(M)
}
\arguments{
\item{replist}{List object created by read_admb function}
}
\value{
Plot growth increment for given pre-molt size, including model predictions and data
}
\description{
Plot growth from arbitrary start age
}

