% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_growth_transition}
\alias{plot_growth_transition}
\title{Plot growth transition}
\usage{
plot_growth_transition(M, xlab = "Size-class",
  ylab = "P(size transition | molt)")
}
\arguments{
\item{M}{list of object(s) created by read_admb function}

\item{xlab}{the x-axis label for the plot}

\item{ylab}{the y-axis label for the plot}
}
\value{
plot of size transition matrix
}
\description{
The growth probabilities (for all crabs that molt)
}
\author{
SJD Martell, DN Webber
}

