% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-transitions.R
\name{plot_growth_transition}
\alias{plot_growth_transition}
\title{Plot growth transition}
\usage{
plot_growth_transition(M, xlab = "Mid-point of size-class (mm)",
  ylab = "P(growth transition)", slab = "Sex")
}
\arguments{
\item{M}{a list of object(s) created by read_admb function}

\item{xlab}{the x-axis label for the plot}

\item{ylab}{the y-axis label for the plot}

\item{slab}{the sex label for the plot that appears above the key}
}
\value{
plot of size transition matrix
}
\description{
The sex-specific growth transition probabilities (G_h, for all crabs that molt)
}
\author{
SJD Martell, D'Arcy N. Webber
}
