% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_length_weight}
\alias{plot_length_weight}
\title{Plot length-weight relationship}
\usage{
plot_length_weight(M, xlab = "Size", ylab = "Weight")
}
\arguments{
\item{M}{list object created by read_admb function}

\item{xlab}{the x-axis label for the plot}

\item{ylab}{the y-axis label for the plot}
}
\value{
plot of the length-weight relationship
}
\description{
Plot length-weight relationship
}
\author{
DN Webber
}

