% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-length-weight.R
\name{plot_length_weight}
\alias{plot_length_weight}
\title{Plot length-weight relationship}
\usage{
plot_length_weight(M, xlab = "Mid-point of size class (mm)",
  ylab = "Weight (tonnes)")
}
\arguments{
\item{M}{list object created by read_admb function}

\item{xlab}{the x-axis label for the plot}

\item{ylab}{the y-axis label for the plot}
}
\value{
plot of the length-weight relationship
}
\description{
Plot length-weight relationship
}
\author{
DN Webber
}

