% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-natural-mortality.R
\name{plot_natural_mortality}
\alias{plot_natural_mortality}
\title{Plot natural mortality}
\usage{
plot_natural_mortality(M, plt_knots = TRUE, knots = c(1976, 1980, 1985,
  1994), slab = "Knot")
}
\arguments{
\item{M}{list object created by read_admb function}

\item{plt_knots}{if the knots should be plotted or not}
}
\value{
plot of natural mortality
}
\description{
Plot natural mortality
}
\author{
J Ianelli, SJD Martell, DN Webber
}
