% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-numbers.R
\name{plot_numbers}
\alias{plot_numbers}
\title{Plot numbers at length}
\usage{
plot_numbers(M, subsetby = "", nrow = 2, ncol = NULL)
}
\arguments{
\item{M}{list object created by read_admb function}

\item{subsetby}{a selection of the years to restrict plotting of}
}
\value{
plot of numbers of individuals in each size-class each year in the model
}
\description{
Plot numbers at length
}
\author{
D'Arcy N. Webber
}
