% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_recruitment}
\alias{plot_recruitment}
\title{Plot predicted recruitment and approximate asymptotic error-bars}
\usage{
plot_recruitment(M)
}
\arguments{
\item{M}{List object created by read_admb function}
}
\value{
Plot of predicted recruitment
}
\description{
Plot predicted recruitment and approximate asymptotic error-bars
}
\author{
SJD Martell
}

