% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-recruitment.R
\name{plot_recruitment}
\alias{plot_recruitment}
\title{Plot predicted recruitment and approximate asymptotic error-bars}
\usage{
plot_recruitment(M, xlab = "Year",
  ylab = "Recruitment (millions of individuals)")
}
\arguments{
\item{M}{list object created by read_admb function}

\item{xlab}{the x-axis label for the plot}

\item{ylab}{the y-axis label for the plot}
}
\value{
Plot of predicted recruitment
}
\description{
Plot predicted recruitment and approximate asymptotic error-bars
}
\author{
SJD Martell, DN Webber
}
