% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-recruitment.R
\name{plot_recruitment_size}
\alias{plot_recruitment_size}
\title{Plot recruitment size distribution}
\usage{
plot_recruitment_size(M)
}
\arguments{
\item{M}{list object created by read_admb function}
}
\value{
plot of recruitment size distribution
}
\description{
Plot recruitment size distribution
}
\author{
DN Webber
}

