% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-selectivity.R
\name{plot_selectivity}
\alias{plot_selectivity}
\title{Plot selectivity}
\usage{
plot_selectivity(M)
}
\arguments{
\item{replist}{List object created by read_admb function}
}
\value{
Plot of selectivity
}
\description{
Plot selectivity
}
\author{
SJD Martell, DN Webber
}

