% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_selectivity}
\alias{plot_selectivity}
\title{Plot selectivity}
\usage{
plot_selectivity(M)
}
\arguments{
\item{M}{list object created by read_admb function}
}
\value{
plot of selectivity
}
\description{
Plot selectivity
}
\author{
SJD Martell, DN Webber
}

