% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-size-comps.R
\name{plot_size_comps}
\alias{plot_size_comps}
\title{Plot fits to size composition data}
\usage{
plot_size_comps(M, which_plots = "all",
  xlab = "Mid-point of size-class (mm)", ylab = "Proportion",
  slab = "Sex", mlab = "Model", tlab = "Fleet", res = FALSE)
}
\arguments{
\item{M}{List object(s) created by read_admb function}

\item{which_plots}{the size composition fits that you want to plot}

\item{xlab}{the x-axis label for the plot}

\item{ylab}{the y-axis label for the plot}

\item{slab}{the sex label for the plot that appears above the key}

\item{mlab}{the model label for the plot that appears above the key}

\item{tlab}{the fleet label for the plot that appears above the key}

\item{res}{boolean if residual or observed and predicted}
}
\value{
Plots of observed and predicted size composition values
}
\description{
Get observed and predicted size composition values
}

