% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-transitions.R
\name{plot_size_transition}
\alias{plot_size_transition}
\title{Plot size transition}
\usage{
plot_size_transition(M, xlab = "Size-class (mm)",
  ylab = "P(size transition | molt)", slab = "Sex", females = FALSE)
}
\arguments{
\item{M}{list of object(s) created by read_admb function}

\item{xlab}{the x-axis label for the plot}

\item{ylab}{the y-axis label for the plot}

\item{slab}{the sex label for the plot that appears above the key}

\item{females}{logical indicating if the females are to be plotted or not}
}
\value{
plot of size transition matrix
}
\description{
The combination of growth transitions (G_h) and molting probability (P_h)
represents the size transition.
}
\author{
SJD Martell, D'Arcy N. Webber
}

