% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_size_transition}
\alias{plot_size_transition}
\title{Plot size transition}
\usage{
plot_size_transition(M)
}
\arguments{
\item{M}{List of object(s) created by read_admb function}
}
\value{
Plot of size transition matrix
}
\description{
Plot size transition
}
\author{
SJD Martell, DN Webber
}

