% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-transitions.R
\name{plot_size_transition}
\alias{plot_size_transition}
\title{Plot size transition}
\usage{
plot_size_transition(M,
  xlab = "Mid-point of size-class after transition (mm)",
  ylab = "P(growth transition | molt)", slab = "Sex")
}
\arguments{
\item{M}{list of object(s) created by read_admb function}

\item{xlab}{the x-axis label for the plot}

\item{ylab}{the y-axis label for the plot}

\item{slab}{the sex label for the plot that appears above the key}
}
\value{
plot of size transition matrix
}
\description{
The combination of growth transitions (G_h) and molting probability (P_h)
represents the size transition.
}
\author{
SJD Martell, D'Arcy N. Webber
}
