% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_ssb}
\alias{plot_ssb}
\title{Plot predicted spawning stock biomass (ssb)}
\usage{
plot_ssb(M, xlab = "Year", ylab = "MMB (tonnes)")
}
\arguments{
\item{M}{List object(s) created by read_admb function}
}
\value{
Plot of model estimates of spawning stock biomass
}
\description{
Spawning biomass may be defined as all males or some combination of males and
females
}
\author{
SJD Martell, DN Webber
}

