% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_rep}
\alias{read_rep}
\title{Read ADMB .rep file}
\usage{
read_rep(fn)
}
\arguments{
\item{repfile}{name of ADMB output file to be read (no extension needed)}
}
\value{
object of type 'list' with ADMB outputs therein
}
\description{
Read ADMB .rep file and return an R object of type 'list'
}
\author{
Steve Martell
}

