% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_GMACS_dat.R
\name{readGMACS.dat}
\alias{readGMACS.dat}
\title{Read the gmacs.dat file}
\usage{
readGMACS.dat(path = NULL, verbose = TRUE)
}
\arguments{
\item{path}{\itemize{
\item path to the gmacs.dat file
}}

\item{verbose}{\itemize{
\item (TRUE/FALSE); flag to print processing information
}}
}
\value{
the gmacs.dat file as a named list.
}
\description{
Read the gmacs.dat file.
}
\seealso{
\code{\link{readGMACSdat}},\code{\link{readGMACSctl}},
\code{\link{readGMACSprj}}
}
