% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGMACSctl.R
\name{readGMACSctl}
\alias{readGMACSctl}
\title{readGMACSctl}
\usage{
readGMACSctl(FileName = NULL, verbose = TRUE, DatFile = NULL, nyrRetro = NULL)
}
\arguments{
\item{FileName}{\itemize{
\item path (and name (e.g. snow.ctl)) of the control file
}}

\item{verbose}{\itemize{
\item (TRUE/FALSE); flag to print processing information
}}

\item{DatFile}{\itemize{
\item Object (list) containing the .dat file - This is the output
of the \code{\link[=readGMACSdat]{readGMACSdat()}} function.
}}
}
\value{
the .ctl file as a named list.
}
\description{
Read the GMACS control file.
}
\seealso{
\code{\link{readGMACS.dat}},\code{\link{readGMACSdat}},
\code{\link{readGMACSprj}}
}
