% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGMACSdat.R
\name{readGMACSdat}
\alias{readGMACSdat}
\title{Read the GMACS data file}
\usage{
readGMACSdat(FileName = NULL, verbose = TRUE)
}
\arguments{
\item{FileName}{(character string)- path (and name) of the data file (e.g. snow.data)}

\item{verbose}{(logical)- (TRUE/FALSE); flag to print processing information}
}
\value{
the .dat file as a named list.
}
\description{
Read the GMACS data file.
}
\seealso{
\code{\link{readGMACS.dat}},\code{\link{readGMACSctl}},
\code{\link{readGMACSprj}}
}
