% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGMACSrep.R
\name{readGMACSrep}
\alias{readGMACSrep}
\title{Read Gmacs.rep}
\usage{
readGMACSrep(
  FileName = NULL,
  verbose = TRUE,
  DatFile = NULL,
  CtlFile = NULL,
  nyrRetro = NULL
)
}
\arguments{
\item{FileName}{(character string)- path (and name (e.g. gmacs.rep)) of the report file}

\item{verbose}{(logical)- (TRUE/FALSE); flag to print processing information}

\item{DatFile}{(list)- Object containing the .dat file - This is the output
of the \code{\link[=readGMACSdat]{readGMACSdat()}} function.}

\item{CtlFile}{(list)- Object containing the .ctl file - This is the output
of the \code{\link[=readGMACSctl]{readGMACSctl()}} function.}

\item{nyrRetro}{(integer)- Number of year for the retrospective analysis}
}
\value{
the gmacs.rep file as a named list.
}
\description{
This functions reads the gmacs.rep file (report file) as a named list
within R.
}
\seealso{
\code{\link{readGMACSdat}}, \code{\link{readGMACSctl}}, \code{\link{readGMACSallOUT}}
}
