% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGMACSsimdat.R
\name{readGMACSsimdat}
\alias{readGMACSsimdat}
\title{readGMACSsimdat}
\usage{
readGMACSsimdat(
  Dir = NULL,
  FileName = "simdata.out",
  verbose = TRUE,
  DatFile = NULL,
  CtlFile = NULL
)
}
\arguments{
\item{Dir}{(character string)- path to the "simulated" data file (output from GMACS)}

\item{FileName}{(character string)- name of the "simulated" data file - By default, "simdata.out"}

\item{verbose}{(logical)- flag to print processing information}

\item{DatFile}{(list)- Object containing the .dat file - This is the output
of the \code{\link[=readGMACSdat]{readGMACSdat()}} function.}

\item{CtlFile}{(list)- Object (list) containing the .ctl file - This is the output
of the \code{\link[=readGMACSctl]{readGMACSctl()}} function.}
}
\value{
the simdata.out file as a named list.
}
\description{
Read the simdata.out file. This is an output of GMACS that holds
the predicted catch and cpue from an assessment run as well as the observed
and predicted size compositions.
}
\seealso{
\code{\link{readGMACS.dat}},\code{\link{readGMACSdat}},\code{\link{readGMACSctl}},
\code{\link{readGMACSprj}}, \code{\link{readGMACSpar}}
}
