% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-likelihoods.R
\name{table_penalties}
\alias{table_penalties}
\title{Create a table of penalties}
\usage{
table_penalties(M)
}
\arguments{
\item{M}{list object(s) created by read_admb function}
}
\value{
dataframe of likelihood components for each model in M
}
\description{
Nees to be fixed if Jim changes order
}
\author{
D'Arcy N. Webber
}
