% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeGmacsPAR.R
\name{writeGmacsPAR}
\alias{writeGmacsPAR}
\title{Write the gmacs.par file}
\usage{
writeGmacsPAR(Dir = NULL, FileName = NULL, gmacsPar = NULL)
}
\arguments{
\item{Dir}{(character string)- path where to save the new gmacs.par file}

\item{FileName}{(character string)- name of the new gmacs.par file}

\item{gmacsPar}{(character string)- Object (list) containing the ex gmacs.par file - The list is
created using the \code{\link[=readGMACSpar]{readGMACSpar()}} function.}
}
\value{
create a new gmacs.dat file.
}
\description{
Write a new gmacs.par file. This function is used to modify within
R a pre-existent gmacs.par file.
}
\seealso{
\code{\link{readGMACSpar}}
}
